vdo_github/requirements.txt
PyGithub>=2.1.0
GitPython>=3.1.40
python-dotenv>=1.0.0
